Add bullet object to game.
Bullet object was added, where states are intialized at the same location as the player via the board.
We still need to determine how the y coordinate will change over time and how to best handle that.

The next steps will be to set up methods for how objects will move.