Refactored code for collision detection.
Added two board prototype functions to check and remove objects when a collision has occured. Tests were also deleted and updated
to reflect changes. Tests for the object ranges and meteor collision check were deleted, since those functions were deleted.
This improves collision detection, as it can now be performed on any set of objects and is not limited to the largest object as before.

The next step will be to refacotr collision detection code.