Completed collision detection.
Added and active state to each object, with the exception of the board, and added lodash to shorten code.
Collisions are detected for meteors and bullets first, then player and meteors are checked and meteors array is
only iterated through a second time when there is a collision with a player. Downside is that bullets and meteors
collections are iterated through everytimet the removeInActiveObjects function is called.

Next step will be to add unit tests for other object statuses.